ProjectManager, FileManager, and Error messages
Added error and warning messages throughout the program to support
future debugging. Some errors will be fatal and will throw IOExceptions.
These are only in cases where it would be pointless for the program to
continue running with those errors.

ProjectManager has been created to handle the reading and writing of
.proj files, as well as handling the creation of a HashMap of Project
objects at startup. The class will automatically search through the
directory to find any existing .proj files and will load them into the
HashMap before the GUI launches.

The FileManager has the simple task of making sure the program has its
directory in order. It throws fatal IOExceptions if it is unable to make
the file system. (Duh.) It should likely only be needed during the first
time the program ever runs.