Image Importing and Viewing:
Edited ImportDialog to take an image selected by JFileChooser and write
a copy to src/data/images. This now allows us to copy images from
anywhere on the disk, and save them to the program's folder.

I also updated the editor panel to both show images, and update them
when the next and previous buttons are pressed, allowing the images of a
project to be cycled between.