Project Panel Buttons
Updated the InnerProjectsPanel to display each project with its
thumbnail on a button, which when clicked opens the project in the
editor. I'm fixed a number of issues with the Project system which only
became apparent now, as there was no way to test them before.

I'm also pushing the images Liz used for the Icons update, since we've
been having issues getting them into the repo.