Selecting Images and Animation Update
Images can now be selected from the existing images in the /images
directory. A new class SelectDialog has been made, which is similar to
ImportDialog but only deals in pathnames to files instead of the actual
files. The other updates are changes to FileManager, the EditorPanel,
MainFrame, and NewProjectPanel - all to support SelectDialog.

I have also added functionality to the Frames/Second component of the
editor, and the play button, by putting in a JTextField to act as the
user input for the frames/sec option. The Play/Pause button works,
starting and stopping a timer which updates and repaints to image in the
EditorPanel.

Finally, I added the import ability to the Toolbar Import button.