Video Exporting Update
Using a package called Xuggler, and with the packages it's dependent
upon, I've added the ability to export video. It's rather simple, with
clicking the export button in the ToolbarPanel calling an exportVideo
function from VideoManager, a new class I made specifically for this
role. It just steals the fps from the text box on the Editor screen and
passes it along, reading in each image and saving it to the mp4 file.
The bugs were minor, and I fixed one small issue with the image type,
now requiring every image read in to be converted to the right kind of
BufferedImage. This was relatively easy to implement, and I'll put in
further functionality surrounding this shortly.